revise more stuff

+++ Real-Estate-Auction-System-API/Controllers/AdminController.cs
@@ -1,5 +1,7 @@
-﻿using BusinessObjects.Dtos.Request;
+﻿using System.Security.Claims;
+using BusinessObjects.Dtos.Request;
 using BusinessObjects.Dtos.Response;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Services;
 
@@ -17,6 +19,7 @@ public class AdminController : ControllerBase
     }
 
     [HttpGet]
+    [Authorize(Roles = "Admin")]
     public async Task<ActionResult<ListResponseBaseDto<AdminListResponseDto>>> GetAdminsAsync(
         [FromQuery] AdminQuery request)
     {
@@ -28,6 +31,7 @@ public class AdminController : ControllerBase
     }
 
     [HttpPost]
+    [Authorize(Roles = "Admin")]
     public async Task<ActionResult<ResultResponse<AddAdminResponseDto>>> AddAdminAsync(
         [FromBody] AddAdminRequestDto request)
     {
@@ -44,6 +48,7 @@ public class AdminController : ControllerBase
     }
 
     [HttpPut("{id}")]
+    [Authorize(Roles = "Admin")]
     public async Task<ActionResult<UpdateAdminResponseDto>> UpdateAdminAsync([FromRoute] Guid id,
         [FromBody] UpdateAdminRequestDto request)
     {
@@ -55,8 +60,16 @@ public class AdminController : ControllerBase
     }
 
     [HttpDelete("{id}")]
+    [Authorize(Roles = "Admin")]
     public async Task<ActionResult<ResultResponse<DeleteAdminResponseDto>>> DeleteAdminAsync([FromRoute] Guid id)
     {
+        var adminId = Guid.Parse(User.FindFirst(claim => ClaimTypes.NameIdentifier.Equals(claim.Type))?.Value ?? string.Empty);
+
+        if (id == adminId)
+        {
+            return Unauthorized();
+        }
+
         var result = await _adminService.DeleteAdminAsync(id);
         if (result.Status == Status.NotFound) return NotFound(result);
         return Ok(result);
